requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
aiohttp=3.9.3=pypi_0
aiosignal=1.3.1=pypi_0
annotated-types=0.6.0=pypi_0
anyio=4.2.0=pypi_0
asgiref=3.7.2=pyhd8ed1ab_0
async-timeout=4.0.3=pypi_0
attrs=23.2.0=pypi_0
backoff=2.2.1=pypi_0
bcrypt=4.1.2=pypi_0
brotli-python=1.1.0=py310h9e9d8ca_1
build=1.0.3=pypi_0
bzip2=1.0.8=h10d778d_5
ca-certificates=2024.2.2=h8857fd0_0
cachetools=5.3.2=pypi_0
certifi=2023.11.17=pyhd8ed1ab_1
cffi=1.16.0=py310hdca579f_0
charset-normalizer=3.3.2=pypi_0
chroma-hnswlib=0.7.3=pypi_0
chromadb=0.4.22=pypi_0
click=8.1.7=pypi_0
clickhouse-connect=0.7.0=py310hb372a2b_0
coloredlogs=15.0.1=pypi_0
dataclasses-json=0.6.4=pypi_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
django=5.0.1=pyhd8ed1ab_0
django-cors-headers=4.3.1=pypi_0
djangorestframework=3.14.0=pyhd8ed1ab_0
exceptiongroup=1.2.0=pypi_0
fastapi=0.108.0=pypi_0
filelock=3.13.1=pypi_0
flatbuffers=23.5.26=pypi_0
frozenlist=1.4.1=pypi_0
fsspec=2023.12.2=pypi_0
google-auth=2.26.1=pypi_0
googleapis-common-protos=1.62.0=pypi_0
greenlet=3.0.3=pypi_0
grpcio=1.60.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httptools=0.6.1=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.20.2=pypi_0
humanfriendly=10.0=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.11.0=pypi_0
importlib-resources=6.1.1=pypi_0
instructorembedding=1.0.1=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.3.2=pypi_0
jsonpatch=1.33=pypi_0
jsonpointer=2.4=pypi_0
kubernetes=29.0.0=pypi_0
langchain=0.1.5=pypi_0
langchain-community=0.0.17=pypi_0
langchain-core=0.1.18=pypi_0
langchain-openai=0.0.5=pypi_0
langsmith=0.0.86=pypi_0
libblas=3.9.0=21_osx64_openblas
libcblas=3.9.0=21_osx64_openblas
libcxx=16.0.6=hd57cbcb_0
libffi=3.4.2=h0d85af4_5
libgfortran=5.0.0=13_2_0_h97931a8_2
libgfortran5=13.2.0=h2873a65_2
liblapack=3.9.0=21_osx64_openblas
libopenblas=0.3.26=openmp_hfef2a42_0
libsqlite=3.44.2=h92b6c6a_0
libzlib=1.2.13=h8a1eda9_5
llvm-openmp=17.0.6=hb6ac08f_0
lz4=4.3.3=py310hf99a7a4_0
lz4-c=1.9.4=hf0c8a7f_0
markupsafe=2.1.4=pypi_0
marshmallow=3.20.2=pypi_0
mmh3=4.0.1=pypi_0
monotonic=1.6=pypi_0
mpmath=1.3.0=pypi_0
multidict=6.0.5=pypi_0
mypy-extensions=1.0.0=pypi_0
ncurses=6.4=h93d8f39_2
networkx=3.2.1=pypi_0
nltk=3.8.1=pypi_0
numpy=1.26.3=py310h4bfa8fc_0
oauthlib=3.2.2=pypi_0
onnxruntime=1.16.3=pypi_0
openai=1.10.0=pypi_0
openssl=3.2.1=hd75f5a5_0
opentelemetry-api=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-common=1.22.0=pypi_0
opentelemetry-exporter-otlp-proto-grpc=1.22.0=pypi_0
opentelemetry-instrumentation=0.43b0=pypi_0
opentelemetry-instrumentation-asgi=0.43b0=pypi_0
opentelemetry-instrumentation-fastapi=0.43b0=pypi_0
opentelemetry-proto=1.22.0=pypi_0
opentelemetry-sdk=1.22.0=pypi_0
opentelemetry-semantic-conventions=0.43b0=pypi_0
opentelemetry-util-http=0.43b0=pypi_0
overrides=7.4.0=pypi_0
packaging=23.2=pypi_0
pandas=2.2.0=py310h276d7da_0
pillow=10.2.0=pypi_0
pip=23.3.2=pyhd8ed1ab_0
posthog=3.1.0=pypi_0
protobuf=4.25.1=pypi_0
pulsar-client=3.4.0=pypi_0
pyasn1=0.5.1=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pydantic=2.5.3=pypi_0
pydantic-core=2.14.6=pypi_0
pypika=0.48.9=pypi_0
pyproject-hooks=1.0.0=pypi_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.13=h00d2728_1_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python-tzdata=2023.4=pyhd8ed1ab_0
python_abi=3.10=4_cp310
pytz=2023.4=pyhd8ed1ab_0
pyyaml=6.0.1=pypi_0
readline=8.2=h9e318b2_1
regex=2023.12.25=pypi_0
requests=2.31.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rova-client=2.0.4=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.2=pypi_0
scikit-learn=1.4.0=pypi_0
scipy=1.12.0=pypi_0
sentence-transformers=2.2.2=pypi_0
sentencepiece=0.1.99=pypi_0
setuptools=69.0.3=pyhd8ed1ab_0
six=1.16.0=pyh6c4a22f_0
sniffio=1.3.0=pypi_0
sqlalchemy=2.0.25=pypi_0
sqlparse=0.4.4=pyhd8ed1ab_0
starlette=0.32.0.post1=pypi_0
sympy=1.12=pypi_0
tenacity=8.2.3=pypi_0
threadpoolctl=3.2.0=pypi_0
tiktoken=0.5.2=pypi_0
tk=8.6.13=h1abcd95_1
tokenizers=0.15.0=pypi_0
tomli=2.0.1=pypi_0
torch=2.2.0=pypi_0
torchvision=0.17.0=pypi_0
tqdm=4.66.1=pypi_0
transformers=4.37.2=pypi_0
typer=0.9.0=pypi_0
typing-inspect=0.9.0=pypi_0
typing_extensions=4.9.0=pyha770c72_0
tzdata=2023d=h0c530f3_0
urllib3=2.1.0=pypi_0
uvicorn=0.25.0=pypi_0
uvloop=0.19.0=pypi_0
watchfiles=0.21.0=pypi_0
websocket-client=1.7.0=pypi_0
websockets=12.0=pypi_0
wheel=0.42.0=pyhd8ed1ab_0
wrapt=1.16.0=pypi_0
xz=5.2.6=h775f41a_0
yarl=1.9.4=pypi_0
zipp=3.17.0=pypi_0
zstandard=0.22.0=py310hd88f66e_0
zstd=1.5.5=h829000d_0
